refactor: extract TopicWorkspace from /solve
